+++ cleo_plugins/FileSystemOperations/FileSystemOperations.cpp
@@ -1,5 +1,6 @@
 #include "plugin.h"
 #include "CLEO.h"
+#include "FileUtils.h"
 
 #include <set>
 
@@ -9,10 +10,19 @@ using namespace plugin;
 class FileSystemOperations 
 {
 public:
+    static std::set<DWORD> m_hFiles;
     static std::set<HANDLE> m_hFileSearches;
 
     static void WINAPI OnFinalizeScriptObjects()
     {
+        // clean up opened files
+        for (auto handle : m_hFiles)
+        {
+            if (!is_legacy_handle(handle))
+                fclose(convert_handle_to_file(handle));
+        }
+        m_hFiles.clear();
+
         // clean up file searches
         for (auto handle : m_hFileSearches) FindClose(handle);
         m_hFileSearches.clear();
@@ -23,7 +33,21 @@ public:
         auto cleoVer = CLEO_GetVersion();
         if (cleoVer >= CLEO_VERSION)
         {
+            InitializeFileFunc(CLEO_GetGameVersion()); // file utils
+
             //register opcodes
+            CLEO_RegisterOpcode(0x0A9A, opcode_0A9A);
+            CLEO_RegisterOpcode(0x0A9B, opcode_0A9B);
+            CLEO_RegisterOpcode(0x0A9C, opcode_0A9C);
+            CLEO_RegisterOpcode(0x0A9D, opcode_0A9D);
+            CLEO_RegisterOpcode(0x0A9E, opcode_0A9E);
+            CLEO_RegisterOpcode(0x0AD5, opcode_0AD5);
+            CLEO_RegisterOpcode(0x0AD6, opcode_0AD6);
+            CLEO_RegisterOpcode(0x0AD7, opcode_0AD7);
+            CLEO_RegisterOpcode(0x0AD8, opcode_0AD8);
+            CLEO_RegisterOpcode(0x0AD9, opcode_0AD9);
+            CLEO_RegisterOpcode(0x0ADA, opcode_0ADA);
+
             CLEO_RegisterOpcode(0x0AAB, Script_FS_FileExists);
             CLEO_RegisterOpcode(0x0AE4, Script_FS_DirectoryExists);
             CLEO_RegisterOpcode(0x0AE5, Script_FS_CreateDirectory);
@@ -58,6 +82,89 @@ public:
         return path;
     }
 
+    //0A9A=3,%3d% = openfile %1d% mode %2d% // IF and SET
+    static OpcodeResult WINAPI opcode_0A9A(CRunningScript* thread)
+    {
+        auto filename = ReadPathParam(thread);
+
+        char mode[16];
+        auto paramType = CLEO_GetOperandType(thread);
+        if (IsImmInteger(paramType) || IsVariable(paramType))
+        {
+            // integer param (for backward compatibility with CLEO 3)
+            union
+            {
+                DWORD uParam;
+                char strParam[4];
+            } param;
+            param.uParam = CLEO_GetIntOpcodeParam(thread);
+            strcpy(mode, param.strParam);
+        }
+        else
+        {
+            CLEO_ReadStringOpcodeParam(thread, mode, sizeof(mode));
+        }
+
+        // either CLEO 3 or CLEO 4 made a big mistake! (they differ in one major unapparent preference)
+        // lets try to resolve this with a legacy mode
+        bool legacy = CLEO_GetScriptVersion(thread) < CLEO_VER_4_3;
+
+        auto handle = open_file(filename.c_str(), mode, legacy);
+        if (handle == NULL)
+        {
+            CLEO_SetIntOpcodeParam(thread, NULL);
+            CLEO_SetThreadCondResult(thread, false);
+            return OR_CONTINUE;
+        }
+
+        m_hFiles.insert(handle);
+        CLEO_SetIntOpcodeParam(thread, handle);
+        CLEO_SetThreadCondResult(thread, true);
+        return OR_CONTINUE;
+    }
+
+    //0A9B=1,closefile %1d%
+    static OpcodeResult WINAPI opcode_0A9B(CRunningScript* thread)
+    {
+        DWORD handle = CLEO_GetIntOpcodeParam(thread);
+        close_file(handle);
+        m_hFiles.erase(handle);
+        return OR_CONTINUE;
+    }
+
+    //0A9C=2,%2d% = file %1d% size
+    static OpcodeResult WINAPI opcode_0A9C(CRunningScript* thread)
+    {
+        DWORD handle = CLEO_GetIntOpcodeParam(thread);
+        auto size = file_get_size(handle);
+        CLEO_SetIntOpcodeParam(thread, size);
+        return OR_CONTINUE;
+    }
+
+    //0A9D=3,readfile %1d% size %2d% to %3d%
+    static OpcodeResult WINAPI opcode_0A9D(CRunningScript* thread)
+    {
+        DWORD handle = CLEO_GetIntOpcodeParam(thread);
+        DWORD size = CLEO_GetIntOpcodeParam(thread);
+        SCRIPT_VAR* buf = CLEO_GetPointerToScriptVariable(thread);
+
+        buf->dwParam = 0; // https://github.com/cleolibrary/CLEO4/issues/91
+        read_file(buf, size, 1, handle);
+        return OR_CONTINUE;
+    }
+
+    //0A9E=3,writefile %1d% size %2d% from %3d%
+    static OpcodeResult WINAPI opcode_0A9E(CRunningScript* thread)
+    {
+        DWORD handle = CLEO_GetIntOpcodeParam(thread);
+        DWORD size = CLEO_GetIntOpcodeParam(thread);
+        SCRIPT_VAR* buf = CLEO_GetPointerToScriptVariable(thread);
+
+        write_file(buf, size, 1, handle);
+        flush_file(handle);
+        return OR_CONTINUE;
+    }
+
     // 0AAB=1, file_exists %1s%
     static OpcodeResult WINAPI Script_FS_FileExists(CRunningScript* thread)
     {
@@ -67,6 +174,169 @@ public:
         bool exists = (fAttr != INVALID_FILE_ATTRIBUTES) && !(fAttr & FILE_ATTRIBUTE_DIRECTORY);
 
         CLEO_SetThreadCondResult(thread, exists);
+        return OR_CONTINUE;
+    }
+
+    //0AD5=3,file %1d% seek %2d% from_origin %3d% //IF and SET
+    static OpcodeResult WINAPI opcode_0AD5(CRunningScript* thread)
+    {
+        DWORD handle = CLEO_GetIntOpcodeParam(thread);
+        int offset = (int)CLEO_GetIntOpcodeParam(thread);
+        int origin = (int)CLEO_GetIntOpcodeParam(thread);
+
+        FILE* file = convert_handle_to_file(handle);
+        if (file == nullptr)
+        {
+            CLEO_SetThreadCondResult(thread, false);
+            return OR_CONTINUE;
+        }
+
+        bool ok = fseek(file, offset, origin) == 0;
+        CLEO_SetThreadCondResult(thread, ok);
+        return OR_CONTINUE;
+    }
+
+    //0AD6=1,end_of_file %1d% reached
+    static OpcodeResult WINAPI opcode_0AD6(CRunningScript* thread)
+    {
+        DWORD handle = CLEO_GetIntOpcodeParam(thread);
+
+        FILE* file = convert_handle_to_file(handle);
+        if (file == nullptr)
+        {
+            CLEO_SetThreadCondResult(thread, true);
+            return OR_CONTINUE;
+        }
+
+        bool end = ferror(file) || feof(file) != 0;
+        CLEO_SetThreadCondResult(thread, end);
+        return OR_CONTINUE;
+    }
+
+    //0AD7=3,read_string_from_file %1d% to %2d% size %3d% //IF and SET
+    static OpcodeResult WINAPI opcode_0AD7(CRunningScript* thread)
+    {
+        DWORD handle = CLEO_GetIntOpcodeParam(thread);
+
+        char* buffer = nullptr;
+        int bufferSize = 0;
+        DWORD needsTerminator = TRUE;
+        CLEO_ReadStringParamWriteBuffer(thread, &buffer, &bufferSize, &needsTerminator);
+
+        int size = CLEO_GetIntOpcodeParam(thread);
+
+        if (size <= 0)
+        {
+            CLEO_SetThreadCondResult(thread, false);
+            return OR_CONTINUE;
+        }
+
+        FILE* file = convert_handle_to_file(handle);
+        if (file == nullptr)
+        {
+            CLEO_SetThreadCondResult(thread, false);
+            return OR_CONTINUE;
+        }
+
+        std::vector<char> tmpBuff;
+        tmpBuff.resize(size);
+        auto data = tmpBuff.data();
+
+        bool ok = fgets(data, size, file) != nullptr;
+        if(!ok)
+        {
+            CLEO_SetThreadCondResult(thread, false);
+            return OR_CONTINUE;
+        }
+
+        // copy into result param
+        int len = strlen(data);
+        int resultSize = min(len, bufferSize - (int)needsTerminator);
+
+        memcpy(buffer, data, resultSize);
+        if(resultSize < bufferSize) buffer[resultSize] = '\0'; // terminate string whenever possible
+
+        else CLEO_SetThreadCondResult(thread, false);
+        return OR_CONTINUE;
+    }
+
+    //0AD8=2,write_string_to_file %1d% from %2d% //IF and SET
+    static OpcodeResult WINAPI opcode_0AD8(CRunningScript* thread)
+    {
+        DWORD handle = CLEO_GetIntOpcodeParam(thread);
+        auto text = CLEO_ReadStringOpcodeParam(thread);
+
+        FILE* file = convert_handle_to_file(handle);
+        if (file == nullptr)
+        {
+            CLEO_SetThreadCondResult(thread, false);
+            return OR_CONTINUE;
+        }
+
+        bool ok = fputs(text, file) > 0;
+        if (!ok)
+        {
+            CLEO_SetThreadCondResult(thread, false);
+            return OR_CONTINUE;
+        }
+
+        fflush(file);
+        CLEO_SetThreadCondResult(thread, true);
+        return OR_CONTINUE;
+    }
+
+    //0AD9=-1,write_formated_text %2d% to_file %1d%
+    static OpcodeResult WINAPI opcode_0AD9(CRunningScript* thread)
+    {
+        DWORD handle = CLEO_GetIntOpcodeParam(thread);
+        auto format = CLEO_ReadStringOpcodeParam(thread);
+        char text[MAX_STR_LEN]; CLEO_ReadParamsFormatted(thread, format, text, MAX_STR_LEN);
+        
+        FILE* file = convert_handle_to_file(handle);
+        if (file == nullptr)
+        {
+            return OR_CONTINUE;
+        }
+        
+        fputs(text, file);
+        fflush(file);
+        return OR_CONTINUE;
+    }
+
+    //0ADA=-1,%3d% = scan_file %1d% format %2d% //IF and SET
+    static OpcodeResult WINAPI opcode_0ADA(CRunningScript* thread)
+    {
+        DWORD handle = CLEO_GetIntOpcodeParam(thread);
+        auto format = CLEO_ReadStringOpcodeParam(thread);
+        int* result = (int*)CLEO_GetPointerToScriptVariable(thread);
+
+        size_t paramCount = 0;
+        SCRIPT_VAR* ExParams[35];
+        while (CLEO_GetOperandType(thread) != eDataType::DT_END)
+        {
+            // TODO: if target param is string variable it should be handled correctly
+            ExParams[paramCount++] = CLEO_GetPointerToScriptVariable(thread);
+        }
+        CLEO_SkipUnusedVarArgs(thread); // var arg terminator
+
+        FILE* file = convert_handle_to_file(handle);
+        if (file == nullptr)
+        {
+            *result = 0;
+            CLEO_SetThreadCondResult(thread, false);
+            return OR_CONTINUE;
+        }
+
+        *result = fscanf(file, format,
+            /* extra parameters (will be aligned automatically, but the limit of 35 elements maximum exists) */
+            ExParams[0], ExParams[1], ExParams[2], ExParams[3], ExParams[4], ExParams[5],
+            ExParams[6], ExParams[7], ExParams[8], ExParams[9], ExParams[10], ExParams[11],
+            ExParams[12], ExParams[13], ExParams[14], ExParams[15], ExParams[16], ExParams[17],
+            ExParams[18], ExParams[19], ExParams[20], ExParams[21], ExParams[22], ExParams[23],
+            ExParams[24], ExParams[25], ExParams[26], ExParams[27], ExParams[28], ExParams[29],
+            ExParams[30], ExParams[31], ExParams[32], ExParams[33], ExParams[34]);
+
+        CLEO_SetThreadCondResult(thread, paramCount == *result);
         return OR_CONTINUE;
     }
 
@@ -333,4 +603,5 @@ public:
     }
 } fileSystemOperations;
 
+std::set<DWORD> FileSystemOperations::m_hFiles;
 std::set<HANDLE> FileSystemOperations::m_hFileSearches;